Large reconfiguration
Broke the single file into structure agreed on
for extensibility
Also added configuration file to gitignore; this stops people from
accidentally submitting passwords
