+++ .gitignore
@@ -102,3 +102,4 @@ ENV/
 
 # Pycharm
 .idea/*
+/pyeo/conf.ini
