+++ pyeo/conf.ini
@@ -1,6 +1,6 @@
 [sen2]
-user= 
-pass = 
+user = jfr10
+pass = gabomba58
 
 [query]
 #aoi_format: can be shapefile or geojson
