--- pyeo/tests/test_sen_2_tools.py
@@ -1,4 +1,0 @@
-import sen_2_tools
-import tempfile
-
-
