+++ setup.py
@@ -2,7 +2,7 @@ from distutils.core import setup
 
 setup(
     name='Pyeo',
-    version='0.1.0',
+    version='0.2.0',
     author='John Roberts',
     author_email='jfr10@le.ac.uk',
     packages=['pyeo', 'pyeo.tests'],
